refactor: QueryBuilder 패키지 이동
