+++ src/main/java/persistence/sql/ddl/CreateQueryBuilder.java
@@ -5,6 +5,7 @@ import jakarta.persistence.GeneratedValue;
 import jakarta.persistence.GenerationType;
 import jakarta.persistence.Id;
 import jakarta.persistence.Transient;
+import persistence.sql.QueryBuilder;
 
 import java.lang.reflect.Field;
 import java.util.Arrays;
