+++ src/main/java/persistence/sql/ddl/DropQueryBuilder.java
@@ -1,5 +1,7 @@
 package persistence.sql.ddl;
 
+import persistence.sql.QueryBuilder;
+
 public class DropQueryBuilder extends QueryBuilder {
     private static final String DROP_QUERY_TEMPLATE = "DROP TABLE IF EXISTS %s";
